@@ -8,7 +8,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.18</version>
+        <version>3.3.5</version>
         <relativePath/>
     </parent>
 
@@ -19,16 +19,18 @@
     <description>国资国企数据采集上报系统</description>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <mybatis-plus.version>3.5.5</mybatis-plus.version>
-        <dynamic-datasource.version>4.2.0</dynamic-datasource.version>
-        <druid.version>1.2.20</druid.version>
+        <dynamic-datasource.version>4.3.1</dynamic-datasource.version>
+        <druid.version>1.2.23</druid.version>
         <hutool.version>5.8.25</hutool.version>
         <fastjson.version>2.0.43</fastjson.version>
         <poi.version>5.2.5</poi.version>
         <postgresql.version>42.7.1</postgresql.version>
         <mysql.version>8.0.33</mysql.version>
+        <dm.version>8.1.3.62</dm.version>
+        <redisson.version>3.36.0</redisson.version>
     </properties>
 
     <dependencies>
@@ -85,17 +87,26 @@
             <version>${mysql.version}</version>
         </dependency>
 
+        <!-- DM (达梦) Driver -->
+        <dependency>
+            <groupId>com.dameng</groupId>
+            <artifactId>DmJdbcDriver18</artifactId>
+            <version>${dm.version}</version>
+        </dependency>
+
         <!-- Redis -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-data-redis</artifactId>
+            <optional>true</optional>
         </dependency>
 
         <!-- Redisson (分布式锁) -->
         <dependency>
             <groupId>org.redisson</groupId>
             <artifactId>redisson-spring-boot-starter</artifactId>
-            <version>3.25.2</version>
+            <version>${redisson.version}</version>
+            <optional>true</optional>
         </dependency>
 
         <!-- Hutool 工具类 -->
